+++ src/Interpreters/parseColumnsListForTableFunction.cpp
@@ -19,6 +19,7 @@ namespace Setting
 {
     extern const SettingsBool enable_time_time64_type;
     extern const SettingsBool allow_experimental_nullable_tuple_type;
+    extern const SettingsBool allow_experimental_nullable_array_type;
     extern const SettingsBool allow_suspicious_fixed_string_types;
     extern const SettingsBool allow_suspicious_low_cardinality_types;
     extern const SettingsBool allow_suspicious_variant_types;
@@ -43,6 +44,7 @@ DataTypeValidationSettings::DataTypeValidationSettings(const DB::Settings & sett
     , validate_nested_types(settings[Setting::validate_experimental_and_suspicious_types_inside_nested_types])
     , enable_time_time64_type(settings[Setting::enable_time_time64_type])
     , allow_experimental_nullable_tuple_type(settings[Setting::allow_experimental_nullable_tuple_type])
+    , allow_experimental_nullable_array_type(settings[Setting::allow_experimental_nullable_array_type])
 {
 }
 
@@ -149,6 +151,21 @@ void validateDataType(const DataTypePtr & type_to_check, const DataTypeValidatio
                         ErrorCodes::ILLEGAL_COLUMN,
                         "Cannot create column with type '{}' because Nullable Tuple type is not allowed. "
                         "Set setting allow_experimental_nullable_tuple_type = 1 in order to allow it",
+                        data_type.getName());
+                }
+            }
+        }
+
+        if (!settings.allow_experimental_nullable_array_type)
+        {
+            if (const auto * nullable_type = typeid_cast<const DataTypeNullable *>(&data_type))
+            {
+                if (isArray(nullable_type->getNestedType()))
+                {
+                    throw Exception(
+                        ErrorCodes::ILLEGAL_COLUMN,
+                        "Cannot create column with type '{}' because Nullable Array type is not allowed. "
+                        "Set setting allow_experimental_nullable_array_type = 1 in order to allow it",
                         data_type.getName());
                 }
             }