+++ src/Processors/Formats/Impl/NativeORCBlockInputFormat.cpp
@@ -1716,7 +1716,7 @@ ColumnWithTypeAndName ORCColumnToCHColumn::readColumnFromORCColumn(
     bool skipped = false;
 
     if (!inside_nullable && (orc_column->hasNulls || (type_hint && isNullableOrLowCardinalityNullable(type_hint))) && !orc_column->isEncoded
-        && (orc_type->getKind() != orc::LIST && orc_type->getKind() != orc::MAP))
+        && orc_type->getKind() != orc::MAP)
     {
         DataTypePtr nested_type_hint;
         if (type_hint)